add safety checker to text_to_image.py
