+++ examples/stable_diffusion_v2/tools/safety_checker/safety_checker_2.py
@@ -1,5 +1,11 @@
 import argparse
+import os
+import sys
 
+# add current working dir to path to prevent ModuleNotFoundError
+sys.path.insert(0, os.getcwd())
+
+from tools.safety_checker.safety_checker import SafetyChecker
 from tools._common import L2_norm_ops, load_images
 from tools._common.clip import CLIPImageProcessor, parse
 
@@ -7,7 +13,7 @@ import mindspore as ms
 from mindspore import load_checkpoint, load_param_into_net
 
 
-class SafetyChecker2:
+class SafetyChecker2(SafetyChecker):
     def __init__(
         self,
         backend="ms",
@@ -15,6 +21,7 @@ class SafetyChecker2:
         ckpt_path=None,
         model_name="openai/clip-vit-large-patch14",
         check_certificate=False,
+        threshold=0.2,
         **kwargs,
     ):
         if backend == "pt":
@@ -51,7 +58,8 @@ class SafetyChecker2:
             from tools._common.clip import CLIPModel
 
             config = parse(config, ckpt_path)
-
+            self.image_size = config.vision_config.image_size
+            self.dtype = ms.float32 if config.dtype == 'float32' else ms.float16
             model = CLIPModel(config)
             processor = CLIPImageProcessor()
 
@@ -64,6 +72,7 @@ class SafetyChecker2:
         self.nsfw_model = nsfw_model
         self.processor = processor
         self.backend = backend
+        self.threshold = threshold
 
     def eval(self, paths):
         images, paths = load_images(paths)
@@ -77,6 +86,21 @@ class SafetyChecker2:
         return self.__call__(images)
 
     def __call__(self, images):
+        original_images = images
+
+        if self.backend == 'ms' and (images.shape[-1] != self.image_size or images.shape[-2] != self.image_size):
+            from PIL import Image
+            import numpy as np
+            from mindspore import ops
+
+            images_ = []
+            for i in range(images.shape[0]):
+                im = Image.fromarray((255.0 * images[i].transpose((1, 2, 0))).astype(np.uint8).asnumpy())
+                im = im.resize((self.image_size, self.image_size))
+                im = ms.Tensor(np.asarray(im), self.dtype)
+                images_.append(im)
+            images = ops.stack(images_).transpose((0, 3, 1, 2))
+
         image_features = self.model.get_image_features(images)
         if self.backend == "ms":
             norm = L2_norm_ops(image_features)
@@ -85,7 +109,25 @@ class SafetyChecker2:
         image_features = image_features / norm
 
         scores = self.nsfw_model(image_features)
-        return scores
+
+        has_nsfw_concepts = [score if score > self.threshold else 0 for score in scores]
+
+        if self.backend == 'pt':
+            import torch
+        for idx, has_nsfw_concepts in enumerate(has_nsfw_concepts):
+            if has_nsfw_concepts:
+                if self.backend == 'pt':
+                    original_images[idx] = torch.zeros_like(original_images[idx])
+                else:
+                    original_images[idx] = ops.zeros(original_images[idx].shape)
+                
+        if any(has_nsfw_concepts):
+            print(
+                "Potential NSFW content was detected in one or more images. A black image will be returned instead."
+                " Try again with a different prompt and/or seed."
+            )
+                        
+        return original_images, has_nsfw_concepts
 
 
 if __name__ == "__main__":
@@ -117,6 +159,12 @@ if __name__ == "__main__":
         help="backend to do CLIP model inference for CLIP score compute. Option: ms, pt." " Default: ms",
     )
     parser.add_argument(
+        "--threshold",
+        default=0.2,
+        type=float,
+        help="a 0-1 scalar-valued threshold above which we believe an image is NSFW" " Default: 0.2",
+    )
+    parser.add_argument(
         "--check_certificate",
         action="store_true",
         help="set this flag to check for certificate for downloads (checks)",
@@ -125,6 +173,6 @@ if __name__ == "__main__":
     checker = SafetyChecker2(**vars(args))
 
     assert args.image_path_or_dir is not None
-    scores = checker.eval(args.image_path_or_dir)
+    _, has_nsfw_concepts = checker.eval(args.image_path_or_dir)
 
-    print(scores)
+    print(has_nsfw_concepts)
